requirements.txt
streamlit
python-dotenv
openai>=1.40.0
pypdf2
python-docx
